requirements.txt
# Astro Runtime includes some pre-installed provider packages:
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
#
# This project uses PostgresHook, which lives in the Postgres provider.
apache-airflow-providers-postgres
pandas>=2.0,<3
streamlit>=1.30,<2
psycopg2-binary>=2.9,<3
